fix work flow 3x

+++ contracts/interfaces/IOfferManager.sol
@@ -1,15 +1,15 @@
 // SPDX-License-Identifier: GPL-3.0
 pragma solidity >=0.7.0 <0.9.0;
 
-import "../libraries/AssetLibrary.sol";
+import "../libraries/PoolLibrary.sol";
 import "../libraries/OfferLibrary.sol";
 import "../libraries/RequestLibrary.sol";
 
 interface IOfferManager {
 
-    function getLendingOffer(uint256 offerId) external view returns(OfferLibrary.LendingOffer memory);
+    function getLendingOffer(uint160 offerId) external view returns(OfferLibrary.LendingOffer memory);
 
-    function getBorrowingOffer(uint256 offerId) external view returns(OfferLibrary.BorrowingOffer memory);
+    function getBorrowingOffer(uint160 offerId) external view returns(OfferLibrary.BorrowingOffer memory);
 
     function getBorrowingRequest(uint256 requestId) external view returns (RequestLibrary.BorrowingRequest memory);
