+++ contracts/libraries/OfferLibrary.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0
 pragma solidity >=0.7.0 <0.9.0;
 
-import "./AssetLibrary.sol";
+import "./PoolLibrary.sol";
 
 library OfferLibrary {
     enum Type {
@@ -10,12 +10,12 @@ library OfferLibrary {
     }
 
     struct LendingOffer {
-        uint256 offerId;
+        uint160 offerId;
         address principalToken;
         uint256 currentPrincipal;
         uint256 initialPrincipal;
         uint256 interest;
-        uint daysToMaturity;
+        uint16 daysToMaturity;
         address[] collateralTokens;
         uint160 expiresAt;
         uint160 createdAt;
@@ -23,7 +23,7 @@ library OfferLibrary {
     }
 
     struct BorrowingOffer {
-        uint256 offerId;
+        uint160 offerId;
         address principalToken;
         address collateralToken;
         uint256 currentCollateral;
@@ -31,7 +31,7 @@ library OfferLibrary {
         uint256 currentPrincipal;
         uint256 initialPrincipal;
         uint256 interest;
-        uint daysToMaturity;
+        uint16 daysToMaturity;
         uint160 expiresAt;
         uint160 createdAt;
         address borrower;
